+++ Sources/DashX/GraphQL/Schema/Objects/UnsubscribeContactResponse.graphql.swift
@@ -1,0 +1,12 @@
+// @generated
+// This file was automatically generated and should not be edited.
+
+@_implementationOnly import ApolloAPI
+
+extension DashXGql.Objects {
+  static let UnsubscribeContactResponse = ApolloAPI.Object(
+    typename: "UnsubscribeContactResponse",
+    implementedInterfaces: [],
+    keyFields: nil
+  )
+}